Enhance UsersController and UsersService with role-based access control, improved response structure, and user management features. Implement guards for user creation, updates, and deletions, ensuring proper permissions are enforced. Add logging for user creation and introduce methods for role validation and user retrieval by email and refresh token.

+++ src/modules/users/users.service.ts
@@ -1,31 +1,49 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  // ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 // import { User } from './interfaces/user.interface';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { v4 as uuidv4 } from 'uuid';
-import * as bcrypt from 'bcrypt';
 import { Repository } from 'typeorm';
+import { Logger } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User as UserEntity } from './models/user.entity';
+import { Role } from '../auth/enums/role.enum';
+import { UpdateUserRefreshTokenDto } from './dto/update-user-refreshtoken.dto';
 // import { NotFoundException } from '@nestjs/common';
+import bcrypt from 'bcrypt';
 
 @Injectable()
 export class UsersService {
   constructor(
     @InjectRepository(UserEntity)
     private userRepository: Repository<UserEntity>,
+    private logger: Logger,
   ) {}
 
+  private readonly ROLE_HIERARCHY = {
+    [Role.ADMIN]: 3,
+    [Role.MANAGER]: 2,
+    [Role.CUSTOMER]: 1,
+    [Role.WAITER]: 1,
+    [Role.COOK]: 1,
+    [Role.CASHIER]: 1,
+  };
+
   async create(createUserDto: CreateUserDto): Promise<UserEntity> {
+    this.logger.log('info', `Creating user ${createUserDto.email}`);
     const user = this.userRepository.create({
       ...createUserDto,
       id: uuidv4(),
-      password: await bcrypt.hash(createUserDto.password, 10),
+      password: bcrypt.hashSync(createUserDto.password, 10),
       createdAt: new Date(),
       updatedAt: new Date(),
       isActive: true,
-      roles: 'user',
-      isAdmin: false,
+      role: createUserDto.role || Role.CUSTOMER,
+      isAdmin: createUserDto.role === Role.ADMIN ? true : false,
       isEmailVerified: false,
     });
     return await this.userRepository.save(user);
@@ -45,7 +63,10 @@ export class UsersService {
     return user;
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto): Promise<UserEntity> {
+  async update(
+    id: string,
+    updateUserDto: UpdateUserDto | UpdateUserRefreshTokenDto,
+  ): Promise<UserEntity> {
     const user = await this.userRepository.findOne({
       where: { id: id },
     });
@@ -63,6 +84,54 @@ export class UsersService {
   }
 
   async remove(id: string): Promise<void> {
+    const user = await this.userRepository.findOne({
+      where: { id: id },
+    });
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
     await this.userRepository.delete(id);
+  }
+
+  async findUserByEmail(email: string): Promise<UserEntity | null> {
+    return await this.userRepository.findOne({
+      where: { email: email },
+    });
+  }
+
+  async findOneByRefreshToken(
+    refreshToken: string,
+  ): Promise<UserEntity | null> {
+    return await this.userRepository.findOne({
+      where: { refreshToken: refreshToken },
+    });
+  }
+
+  canUpdateRole(
+    currentUserRole: Role,
+    targetUserRole: Role,
+    updateRole: Role,
+  ): boolean {
+    const currentUserLevel = this.ROLE_HIERARCHY[currentUserRole];
+    const targetUserLevel = this.ROLE_HIERARCHY[targetUserRole];
+    const updateRoleLevel = this.ROLE_HIERARCHY[updateRole];
+
+    if (currentUserRole == Role.ADMIN) return true;
+    if (currentUserRole == Role.MANAGER && targetUserRole == Role.CUSTOMER)
+      return false;
+    if (targetUserLevel >= currentUserLevel) {
+      return false;
+    }
+    return updateRoleLevel < currentUserLevel;
+  }
+
+  canUpdateOrDeleteUser(currentUserRole: Role, targetUserRole: Role): boolean {
+    const currentUserLevel = this.ROLE_HIERARCHY[currentUserRole];
+    const targetUserLevel = this.ROLE_HIERARCHY[targetUserRole];
+
+    if (currentUserRole == Role.ADMIN) return true;
+    if (currentUserRole == Role.MANAGER && targetUserRole == Role.CUSTOMER)
+      return false;
+    return targetUserLevel < currentUserLevel;
   }
 }
